requirements.txt
fastapi
python-multipart
uvicorn[standard]
jinja2
recipe-scrapers
httpx
deepl
pint
ingredient-parser-nlp